fixed last memory issue

+++ libft/libft/ft_strndup.c
@@ -6,13 +6,13 @@
 /*   By: spentti <spentti@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/31 14:48:21 by spentti           #+#    #+#             */
-/*   Updated: 2020/08/07 16:42:47 by spentti          ###   ########.fr       */
+/*   Updated: 2020/08/12 13:36:25 by spentti          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	*ft_strndup(const char *str, size_t n)
+char	*ft_strndup(const char *str, int n)
 {
 	char	*dup;
 	size_t	i;
@@ -20,7 +20,7 @@ char	*ft_strndup(const char *str, size_t n)
 	i = 0;
 	if (!(dup = ft_strnew(n)))
 		return (NULL);
-	while (str[i] && i < n)
+	while (str[i] && i < (size_t)n)
 	{
 		dup[i] = str[i];
 		i++;
